1、添加electron打包。 2、根据平板进行界面样式调整 3、修改通过按钮叫料叫到的是同一波次bug

+++ sources/v1/server/src/main/java/com/furongsoft/agv/services/SiteService.java
@@ -113,7 +113,7 @@ public class SiteService extends BaseService<SiteDao, Site> {
      * @param type     区域类型
      * @return 区域信息集合
      */
-    public List<AgvAreaModel> selectAreasByParentId(int parentId, Integer type) {
+    public List<AgvAreaModel> selectAreasByParentId(long parentId, Integer type) {
         return agvAreaDao.selectAreaByParentId(parentId, type);
     }
